+++ rummage/lib/gui/settings_dialog.py
@@ -23,6 +23,7 @@ import wx
 import re
 from .settings import Settings
 from .editor_dialog import EditorDialog
+from .file_ext_dialog import FileExtDialog
 from .generic_dialogs import yesno, errormsg
 from .localization import _
 from . import gui
@@ -31,6 +32,9 @@ from .. import util
 
 # We really need a better validator
 BACKUP_VALIDATOR = re.compile(r'^[-_a-zA-Z\d.]+$')
+
+MINIMUM_COL_SIZE = 100
+COLUMN_SAMPLE_SIZE = 100
 
 
 class SettingsDialog(gui.SettingsDialog):
@@ -61,7 +65,7 @@ class SettingsDialog(gui.SettingsDialog):
         ]
         history_records = Settings.get_history_record_count(self.history_types)
         self.history_records_cleared = False
-        mode = Settings.get_regex_mode()
+        self.dc = wx.ClientDC(self.m_filetype_listctrl)
 
         self.editor = Settings.get_editor()
         if isinstance(self.editor, (tuple, list)):
@@ -71,6 +75,7 @@ class SettingsDialog(gui.SettingsDialog):
         self.m_single_checkbox.SetValue(Settings.get_single_instance())
         self.m_history_label.SetLabel(self.RECORDS % history_records)
         self.m_history_clear_button.Enable(history_records > 0)
+        mode = Settings.get_regex_mode()
         self.m_bregex_radio.SetValue(mode == rumcore.BREGEX_MODE)
         self.m_regex_radio.SetValue(mode == rumcore.REGEX_MODE)
         self.m_bre_radio.SetValue(mode == rumcore.BRE_MODE)
@@ -118,6 +123,7 @@ class SettingsDialog(gui.SettingsDialog):
 
         self.m_general_panel.Fit()
         self.m_regex_panel.Fit()
+        self.m_encoding_panel.Fit()
         self.m_editor_panel.Fit()
         self.m_notify_panel.Fit()
         self.m_history_panel.Fit()
@@ -125,9 +131,11 @@ class SettingsDialog(gui.SettingsDialog):
         self.m_settings_notebook.Fit()
         self.m_settings_panel.Fit()
         self.Fit()
-        if self.GetSize()[1] < 500:
-            self.SetSize(wx.Size(500, self.GetSize()[1]))
-        self.SetMinSize(wx.Size(500, self.GetSize()[1]))
+        if self.GetSize()[1] < 550:
+            self.SetSize(wx.Size(550, self.GetSize()[1]))
+        self.SetMinSize(wx.Size(550, self.GetSize()[1]))
+
+        self.size_columns()
 
     def localize(self):
         """Translage strings."""
@@ -181,6 +189,15 @@ class SettingsDialog(gui.SettingsDialog):
         self.CHECK_UPDATES = _("Check updates daily")
         self.PRERELEASES = _("Include pre-releases")
         self.CHECK_NOW = _("Check now")
+        self.ENCODING = _("Encoding")
+        self.CHARDET_CHOICE = [
+            _("Fastest"),
+            _("chardet (pure python)"),
+            _("cchardet (C)")
+        ]
+        self.FILE_TYPE = _("File type")
+        self.EXTENSIONS = _("Extensions")
+        self.SPECIAL = _("Special file types:")
 
     def refresh_localization(self):
         """Localize dialog."""
@@ -188,9 +205,10 @@ class SettingsDialog(gui.SettingsDialog):
         self.SetTitle(self.TITLE)
         self.m_settings_notebook.SetPageText(0, self.GENERAL_TAB)
         self.m_settings_notebook.SetPageText(1, self.REGEX_TAB)
-        self.m_settings_notebook.SetPageText(2, self.EDITOR_TAB)
-        self.m_settings_notebook.SetPageText(3, self.NOTIFICATIONS_TAB)
-        self.m_settings_notebook.SetPageText(4, self.HISTORY_TAB)
+        self.m_settings_notebook.SetPageText(2, self.ENCODING)
+        self.m_settings_notebook.SetPageText(3, self.EDITOR_TAB)
+        self.m_settings_notebook.SetPageText(4, self.NOTIFICATIONS_TAB)
+        self.m_settings_notebook.SetPageText(5, self.HISTORY_TAB)
         self.m_single_checkbox.SetLabel(self.SINGLE_INSTANCE)
         self.m_visual_alert_checkbox.SetLabel(self.NOTIFY_POPUP)
         self.m_audio_alert_checkbox.SetLabel(self.ALERT)
@@ -212,12 +230,57 @@ class SettingsDialog(gui.SettingsDialog):
         self.m_update_checkbox.SetLabel(self.CHECK_UPDATES)
         self.m_prerelease_checkbox.SetLabel(self.PRERELEASES)
         self.m_check_update_button.SetLabel(self.CHECK_NOW)
+        self.m_filetype_label.SetLabel(self.SPECIAL)
+        self.m_filetype_listctrl.InsertColumn(0, self.FILE_TYPE)
+        self.m_filetype_listctrl.InsertColumn(1, self.EXTENSIONS)
+
+        encoding = Settings.get_chardet_mode()
+        cchardet_available = Settings.is_cchardet_available()
+        options = self.CHARDET_CHOICE if cchardet_available else self.CHARDET_CHOICE[:1]
+        for x in options:
+            self.m_encoding_choice.Append(x)
+        self.m_encoding_choice.SetSelection(encoding)
+
+        encoding_ext = Settings.get_encoding_ext()
+        keys = sorted(encoding_ext.keys())
+        idx = 0
+        for key in keys:
+            self.m_filetype_listctrl.InsertItem(idx, key)
+            self.m_filetype_listctrl.SetItem(idx, 1, ', '.join(encoding_ext[key]))
+            idx += 1
+
         self.Fit()
+
+    def size_columns(self):
+        """Size columns."""
+
+        columns = self.m_filetype_listctrl.GetColumnCount()
+        widest_cell = [MINIMUM_COL_SIZE] * columns
+
+        for idx in range(self.m_filetype_listctrl.GetItemCount()):
+            for x in range(columns):
+                text = self.m_filetype_listctrl.GetItemText(idx, x)
+                lw = self.dc.GetFullTextExtent(text)[0]
+                width = lw + 30
+                if width > widest_cell[x]:
+                    widest_cell[x] = width
+        total_size = 0
+        for x in range(len(widest_cell)):
+            total_size += widest_cell[x]
+            self.m_filetype_listctrl.SetColumnWidth(x, widest_cell[x])
+
+        if total_size < self.m_filetype_listctrl.GetSize()[0] - 20:
+            self.SetColumnWidth(1, widest_cell[-1] + self.m_filetype_listctrl.GetSize()[0] - total_size)
 
     def history_cleared(self):
         """Return if history was cleared."""
 
         return self.history_records_cleared
+
+    def on_chardet(self, event):
+        """Handle chardet selection."""
+
+        Settings.set_chardet_mode(self.m_encoding_choice.GetCurrentSelection())
 
     def on_check(self, event):
         """Check updates."""
@@ -352,6 +415,24 @@ class SettingsDialog(gui.SettingsDialog):
         """Close on cancel."""
 
         self.Close()
+
+    def on_dclick(self, event):
+        """Handle double click."""
+
+        pos = event.GetPosition()
+        item = self.m_filetype_listctrl.HitTestSubItem(pos)[0]
+        if item != -1:
+            key = self.m_filetype_listctrl.GetItemText(item, 0)
+            extensions = self.m_filetype_listctrl.GetItemText(item, 1)
+            dlg = FileExtDialog(self, extensions)
+            dlg.ShowModal()
+            value = dlg.extensions
+            dlg.Destroy()
+            if extensions != value:
+                Settings.set_encoding_ext({key: value.split(', ')})
+                self.m_filetype_listctrl.SetItem(item, 1, value)
+                self.size_columns()
+        event.Skip()
 
     def on_change_module(self, event):
         """Change the module."""